+++ src/components/LinkToHome.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+const LinkToHome = () => (
+  <>
+    <Link to="/">キャラクター一覧へ</Link>
+  </>
+)
+
+export default LinkToHome
