+++ src/components/ListItemLink.jsx
@@ -1,0 +1,6 @@
+import { ListItem } from '@material-ui/core'
+import React from 'react'
+
+const ListItemLink = (props) => <ListItem button component="a" {...props} />
+
+export default ListItemLink
