+++ src/components/Value.jsx
@@ -1,0 +1,27 @@
+import { Box } from '@material-ui/core'
+import {
+  FastField,
+  useField,
+} from 'formik'
+import {
+  TextField,
+} from 'formik-material-ui'
+import React from 'react'
+
+const Value = React.memo(({ input, ...props }) => {
+  const [, meta] = useField(props.name)
+  const { value } = meta
+
+  return (
+    input
+      ? <FastField component={TextField} size="small" fullWidth margin="none" variant="outlined" {...props} />
+      : (
+        <Box width="100%" minHeight="20px" fontSize="caption2.fontSize" border={1} {...props}>
+          {value}
+        </Box>
+      )
+  )
+})
+Value.displayName = 'Value'
+
+export default Value
