+++ src/components/index.js
@@ -1,0 +1,16 @@
+export { default as BooleanValue } from './BooleanValue'
+export { default as ComputeValue } from './ComputeValue'
+export { default as DateValue } from './DateValue'
+export { default as DirtyDialog } from './DirtyDialog'
+export { default as ErrorBoundary } from './ErrorBoundary'
+export { default as ImageValue } from './ImageValue'
+export { default as Label } from './Label'
+export { default as Label2 } from './Label2'
+export { default as LinkToHome } from './LinkToHome'
+export { default as ListChars } from './ListChars'
+export { default as ListItemLink } from './ListItemLink'
+export { default as MultiLineValue } from './MultiLineValue'
+export { default as PromptIfDirty } from './PromptIfDirty'
+export { default as SkillAbilityValue } from './SkillAbilityValue'
+export { default as SkillNameValue } from './SkillNameValue'
+export { default as Value } from './Value'
