+++ src/forms/index.js
@@ -1,0 +1,3 @@
+export { default as EditCharForm } from './EditCharForm'
+export { default as NewCharForm } from './NewCharForm'
+export { default as ViewCharForm } from './ViewCharForm'
