+++ src/models/Ability.js
@@ -1,0 +1,11 @@
+class Ability {
+  constructor(name) {
+    this.name = name
+    this.score = 10
+    this.modifier = 0
+    this.temporaryScore = 0
+    this.temporaryModifier = 0
+  }
+}
+
+export default Ability
