+++ src/models/Attack.js
@@ -1,0 +1,14 @@
+class Attack {
+  constructor(attack, attackBonus, critical, weight, type, range, notes, damage) {
+    this.attack = attack
+    this.attackBonus = attackBonus
+    this.critical = critical
+    this.weight = weight
+    this.type = type
+    this.range = range
+    this.notes = notes
+    this.damage = damage
+  }
+}
+
+export default Attack
