+++ src/models/GrappleModifier.js
@@ -1,0 +1,11 @@
+class GrappleModifier {
+  constructor(total, bab, strengthModifier, sizeModifier, miscModifier) {
+    this.total = total
+    this.bab = bab
+    this.strengthModifier = strengthModifier
+    this.sizeModifier = sizeModifier
+    this.miscModifier = miscModifier
+  }
+}
+
+export default GrappleModifier
