+++ src/models/HitPoint.js
@@ -1,0 +1,10 @@
+class HitPoint {
+  constructor(total, damageReduction, current, nonlethalDamage) {
+    this.total = total
+    this.damageReduction = damageReduction
+    this.current = current
+    this.nonlethalDamage = nonlethalDamage
+  }
+}
+
+export default HitPoint
