+++ src/models/Initiative.js
@@ -1,0 +1,9 @@
+class Initiative {
+  constructor(total, dexModifier, miscModifier) {
+    this.total = total
+    this.dexModifier = dexModifier
+    this.miscModifier = miscModifier
+  }
+}
+
+export default Initiative
