+++ src/models/Load.js
@@ -1,0 +1,12 @@
+class Load {
+  constructor(lightLoad, mediumLoad, heavyLoad, liftOverHead, liftOffGround, pushOrDrag) {
+    this.lightLoad = lightLoad
+    this.mediumLoad = mediumLoad
+    this.heavyLoad = heavyLoad
+    this.liftOverHead = liftOverHead
+    this.liftOffGround = liftOffGround
+    this.pushOrDrag = pushOrDrag
+  }
+}
+
+export default Load
