+++ src/models/Money.js
@@ -1,0 +1,10 @@
+class Money {
+  constructor(cp, sp, gp, pp) {
+    this.cp = cp
+    this.sp = sp
+    this.gp = gp
+    this.pp = pp
+  }
+}
+
+export default Money
