+++ src/models/Posession.js
@@ -1,0 +1,8 @@
+class Posession {
+  constructor(item, weight) {
+    this.item = item
+    this.weight = weight
+  }
+}
+
+export default Posession
