+++ src/models/ProtectiveItem.js
@@ -1,0 +1,14 @@
+class ProtectiveItem {
+  constructor(name, acBonus, type, dexMax, checkPenalty, spellFailure, weight, specialProperties) {
+    this.name = name
+    this.acBonus = acBonus
+    this.type = type
+    this.dexMax = dexMax
+    this.checkPenalty = checkPenalty
+    this.spellFailure = spellFailure
+    this.weight = weight
+    this.specialProperties = specialProperties
+  }
+}
+
+export default ProtectiveItem
