+++ src/models/ProtectiveItemsTotal.js
@@ -1,0 +1,12 @@
+class ProtectiveItemsTotal {
+  constructor(acBonus, dexMax, checkPenalty, spellFailure, weight, specialProperties) {
+    this.acBonus = acBonus
+    this.dexMax = dexMax
+    this.checkPenalty = checkPenalty
+    this.spellFailure = spellFailure
+    this.weight = weight
+    this.specialProperties = specialProperties
+  }
+}
+
+export default ProtectiveItemsTotal
