+++ src/pages/EditCharPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { Label, LinkToHome } from '../components'
+import { EditCharForm } from '../forms'
+
+const EditCharPage = () => (
+  <>
+    <Label><LinkToHome /></Label>
+    <EditCharForm />
+  </>
+)
+
+export default EditCharPage
