+++ src/pages/NewCharPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { Label, LinkToHome } from '../components'
+import { NewCharForm } from '../forms'
+
+const NewCharPage = () => (
+  <>
+    <Label><LinkToHome /></Label>
+    <NewCharForm />
+  </>
+)
+
+export default NewCharPage
