+++ src/pages/ViewCharPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { Label, LinkToHome } from '../components'
+import { ViewCharForm } from '../forms'
+
+const ViewCharPage = () => (
+  <>
+    <Label><LinkToHome /></Label>
+    <ViewCharForm />
+  </>
+)
+
+export default ViewCharPage
