+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as EditCharPage } from './EditCharPage'
+export { default as HomePage } from './HomePage'
+export { default as NewCharPage } from './NewCharPage'
+export { default as ViewCharPage } from './ViewCharPage'
